+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and the [Common Changelog](https://common-changelog.org/).
 
+## 2025-04-02
+
+- Using PostgreSQL LISTEN/NOTIFY, multiple instances of CMS Platform will now invalidate caches across instances for real-time content change notifications
+- Separate database connection pools are used for the web application and background jobs to improve performance
+
 ## 2025-01-12
 
 - Fixed calendar event details page when a calendar event is selected